Replace JsonSchema.Registry with Deserializers
Instead of using a "registry" object that's just a dumb container for a list of handler functions, there is now a "plural" object with various methods for dealing with the contained data. This puts the onus on the grouping to handle dispatching and existence checking, rather than on the callsite to correctly access the underlying implementation.

- JsonSchema.Registry should now be Deserializers
- JsonSchema.registerDeserializer should now be Deserializers.register
- Registry().registerDeserializer should now be Deserializers().register
- JsonSchema.defaultRegistry and Registry.defaultInstance should now be Deserializers (or more explicitly Deserializers.defaultInstance)
- The underlying MutableMap is no longer publicly accessible.
- Deserializers exposes new methods: get, has, remove, and deserialize