Change Deserializers.register to Deserializers.add
Seems like a small change, but it sets the tone of the terminology for dealing with deserializer functions, since it's the function that'll be used the most. I'd rather add a deserializer to the group than register a deserializer with a DeserializerRegistry - the latter just feels so much more ugly and Java-y.