Rewrite String.jsonEscape() and String.jsonUnescape().
jsonEscape() was rewritten in a functional programming approach, using the fold() operation. This gets rid of the need to check isEmpty(), reference this[0] and recursively call the method, making it considerably more readable and easier to maintain.

jsonUnescape() was rewritten to improve readability and maintainability as well by using replace() and the convenience method replaceAll().